+++ pip-services4-commons-node/obj/src/data/IChangeable.js
@@ -1,4 +1,0 @@
-"use strict";
-/** @module data */
-Object.defineProperty(exports, "__esModule", { value: true });
-//# sourceMappingURL=IChangeable.js.map