+++ pip-services4-commons-node/obj/src/refer/IUnreferenceable.js
@@ -1,4 +1,0 @@
-"use strict";
-/** @module refer */
-Object.defineProperty(exports, "__esModule", { value: true });
-//# sourceMappingURL=IUnreferenceable.js.map