+++ pip-services4-commons-node/obj/src/run/ICleanable.js
@@ -1,4 +1,0 @@
-"use strict";
-/** @module run */
-Object.defineProperty(exports, "__esModule", { value: true });
-//# sourceMappingURL=ICleanable.js.map