+++ pip-services4-commons-node/obj/test/run/TestClass.d.ts
@@ -1,7 +1,0 @@
-export declare class TestClass {
-    private _value2;
-    constructor(value1: any, value2: any);
-    value1: any;
-    get value2(): any;
-    set value2(value: any);
-}