+++ pip-services4-commons-node/src/index.ts
@@ -1,4 +1,5 @@
 /** @module index */
+export * from './build';
 export * from './commands';
 export * from './config';
 export * from './convert';