+++ pip-services4-components-node/src/index.ts
@@ -1,4 +1,0 @@
-/** @module index */
-export * from './build';
-export * from './info';
-export * from './test';