+++ pip-services4-container-node/obj/examples/DummyFactory.d.ts
@@ -1,5 +1,0 @@
-import { Factory } from 'pip-services4-components-node';
-export declare class DummyFactory extends Factory {
-    private static ControllerDescriptor;
-    constructor();
-}