+++ pip-services4-container-node/src/containers/index.ts
@@ -1,0 +1,9 @@
+/** 
+ * @module containers 
+ * 
+ * Todo: Rewrite the description.
+ * 
+ * @preferred
+ */
+export { Container } from './Container';
+export { ProcessContainer } from './ProcessContainer';