+++ pip-services4-container-node/src/index.ts
@@ -41,6 +41,6 @@
 export * from './build';
 export * from './config';
 export * from './refer';
-
-export { Container } from './Container';
-export { ProcessContainer } from './ProcessContainer';
+export * from './containers';
+export * from './info';
+export * from './test';